Adds Env Var for disabling public channel creation. Resolves #84.
This change introduces an environment variable which can be
set to disable the UI for public channel creation. A separate
variable should be set in the environment of the server to
disable the corresponding API functionality. This change also
reorganizes the UI from a select menu to buttons to accomodate
the disabled state.